requirements.txt
marshmallow>=3
marshmallow-sqlalchemy
flask
flask-restful
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
flask-marshmallow
Flask-JWT==0.2.0
Flask-JWT-Extended==4.2.3
itsdangerous==2.0.1
python-dotenv
celery>=5.0.0
gunicorn
passlib
SQLAlchemy==1.4
psycopg2-binary==2.9.3
flower
py7zr
flask-cors
google-cloud-storage==2.8.0